refactor(Car): Car 클래스를 외부 파일로 분리
