+++ src/App.js
@@ -1,28 +1,5 @@
-import { Console, Random } from '@woowacourse/mission-utils';
-
-class Car {
-  constructor(name) {
-    this.name = name;
-    this.moveCount = 0;
-  }
-
-  getName() {
-    return this.name;
-  }
-
-  forwardOrNot() {
-    let num = Random.pickNumberInRange(0, 9);
-    if (num >= 4) this.moveCount++;
-  }
-
-  getProgress() {
-    Console.print(`${this.name} : ${'-'.repeat(this.moveCount)}`);
-  }
-
-  getCount() {
-    return this.moveCount;
-  }
-}
+import Console from '@woowacourse/mission-utils';
+import Car from './Car.js';
 
 async function getNameInput() {
   const input = Console.readLineAsync(
